TC-11 Create function createAnAccount and Update file name
